pyproject.toml
[project]
name = "cycling-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosql>=15.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.129.0",
    "httpx[http2]>=0.28.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "rapidfuzz>=3.14.3",
    "yoyo-migrations>=9.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=5.0.0",
]
